backend/requirements.txt
# Flask & config
flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.1

# Transformers
transformers==4.39.3

# Torch stack
torch==2.2.2
torchvision==0.17.2

# Image & similarity tools
Pillow==10.2.0
scikit-learn==1.3.2
numpy==1.26.4  # Compatible with torch 2.2.x

# Optional debugging & utilities
requests==2.31.0
